Update patch set 2
Patch Set 2:

> > As for standards, I think the whole idea of casting to char, doing pointer math and casting back is not really allowed, but I don't think it's going to create any problems either?
> 
> Good question. I think that's also undefined behaviour -- and I'm not sure which the optimiser would be more forgiving of. That's the problem with writing an OS kernel in C++ instead of C. A lot more behaviour is undefined.

Well, the typical use case will be:
- 1) We allocate a buffer
- 2) We read some data to it (from network, file, etc)
- 3) We cast the pointer to some type to parse the content
- 4) We cast back to char from time to time to increment the pointer
- 5) and then cast to a specific type again

3 and 5 assume we can cast to the upper type and then dereference. This requires the alignment to be correct. reinterpret_cast<> indeed explicitly allows do cast between char and other types for pointers, and the aliasing rule holds, but it makes no check about the alignment. The aliasing rules is not that important for the most part, because we never access the data as char directly, except when initially filling the buffer.

On a platform where alignment is required, this would indeed be UB (even in C), for example old ARM CPUs would just rotate the bytes of a single aligned word when doing misaligned access. But there is no way around this, as there is no standard way to check if a pointer is indeed aligned properly before dereferencing.

We will have to rely on compilers doing something sane on the architectures we plan to support here...

Patch-set: 2