Update patch set 2
Patch Set 2:

Yeah, the creator of the package kit named these functions quite weirdly...Basically, LatestActivePackageInfos() returns the packages on the system that are both installed and fully set up. When packages are in the middle of being installed, LatestInactivePackageInfos() shows the packages in the process of being installed. Once the installation process is done, LatestInactivePackageInfos() returns nothing. If there are packages that can't be fully activated without a reboot, CurrentlyActivePackageInfos() will return the same information as LatestActivePackageInfos(), or if everything has been installed and activated, it will return no packages. Quite a weird naming/functionality mess, though at least through testing, this new method should work as intended.

Patch-set: 2
Label: Code-Review=0