Update patch set 2
Patch Set 2: Code-Review-2

The font size is a float, not an integer. So, using a shift does not work.

Even if it was an integer, it is better to use multiplications and divisions. Micro-optimisations like these should be left to the compiler, and we should keep the code readable for humans, that's more important.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2