Implement public suffix list to filter cookies
It should not be possible to set a cookie on a public suffix
 (eg *.github.io, *.co.uk, ...).
Currently it is not filtered.This allows users to set cookies
that affect other website sharing the same suffix,
allowing them tp spy on and track users.
Libpsl is a C library that handles public suffix list.
we have to integrate the library into our buildsystem.
it is available on haikuports :
https://github.com/haikuports/haikuports/blob/c841c3f2206c4bcf240bb774570eaa5f5d17451b/net-libs/libpsl/libpsl-0.20.2.recipe
I have added it to the x86_gcc2 and x86_64 package list.
Plesae put the package on the server.
Fixes: #14810