Update patch set 1
Patch Set 1:

Next time you want to update a change, you can include the "change id" in the commit message. It is available in the review page and looks like this:

Change-Id: I44f038e62f0384613063ab694b2200dc514f7b2c

This allows Gerrit to identify that your change is an update to a previous one, instead of a new change. It makes it easier to keep all the comments and discussions in one place.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0