Update patch set 1
Patch Set 1:

> Patch Set 1:
> 1. this is expected. preinit functions don't exist in shared libraries, no need to call them.

Except our applications are often also shared libraries and vice versa; in fact the Network preferences panel is treated as a "shared object" when linking some of the add-ons. Also due to replicants, this situation may be reversed in certain cases. So we cannot assume that only the main application binary has preinit functions; we need to call preinit functions in everything, right?

> 2. Hmm I don't see where we might initialize twice. Please clarify.

get_sorted_image_list with flag argument returns only those libraries which are not initialized; and then initHead will ignore (or not) the head element. We need to take all this into account.

Patch-set: 1