Update patch set 1
Patch Set 1:

(4 comments)

I am not sure B_UNDERSCORE_FACE is the way to go at all. The most important reason being that it doesn't allow for underlines in a different color than the text itself. The next problem is that there are different kinds of underlines (line styles: single, double, dotted, wiggle,...). This whole approach is not flexible enough. For applications, it is simple enough to underline text. Where it gets interesting is when the application doesn't want to draw over glyphs extending below the baseline. So some new drawing API would be nice to draw "something" near text, but with excluding of glyph shapes. Something like BView::ClipToInverseText(const char* ...), and then any drawing operations after that will not overlap the text passed, until clipping is reset. The code to implement this will be a bit involved, but there is already code for setting an alpha mask (which I think ClipToPicture() is using), then the new ClipToInverseText() could use that. Perhaps it could be implemented entirely on the client side, if a faux-bold feature is added to text rendering. This needs to be done in BFont then (faux-bold-width becomes a new property of BFont).
OK, enough rambling. I hope I have provided some food for thought... :-) What do others think?

Patch-set: 1