Update patch set 1
Patch Set 1:

> Where it gets interesting is when the application doesn't want to draw over glyphs extending below the baseline. 

Wouldn't you just draw the underline (or other decorations) before the text, then?
I'm not sure we can get it to look correct with antialiasing and subpixel, otherwise.

Anyway, I think we could still have B_UNDERSCORE_FACE or something similarly very simple for underlines (they are quite common for example for hyperlinks). It is also not as simple as it sounds with linebreaks, etc. I feel it would be quite annoying to have the application manage this manually, eg. in a word processor, and it would be great if the style could be fully embedded in text_run arrays as all other style elements.

I see the need for more underline styles, however, but I don't think letting the Application side handle it all is a great compromise.

Patch-set: 1