Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Where it gets interesting is when the application doesn't want to draw over glyphs extending below the baseline. 
> 
> Wouldn't you just draw the underline (or other decorations) before the text, then?
> I'm not sure we can get it to look correct with antialiasing and subpixel, otherwise.

The underline is supposed to have wider gaps, it shouldn't look simply as if the line is drawn first and then the glyph on top. MacOS supports this, it you want to take a look.

 #####
##  ##
##  ##
 #####
--  ## --
##  ##
 ####

versus

 #####
##  ##
##  ##
 #####
----##---
##  ##
 ####


> Anyway, I think we could still have B_UNDERSCORE_FACE or something similarly very simple for underlines (they are quite common for example for hyperlinks). It is also not as simple as it sounds with linebreaks, etc. I feel it would be quite annoying to have the application manage this manually, eg. in a word processor, and it would be great if the style could be fully embedded in text_run arrays as all other style elements.

OK, I agree. A simple case for a single underline in the same color as the text. However, the patch as is would require a special font file which includes the underline. I think that is not how it will be actually working, although if such a font exists, it could be wise to use it instead of the manual implementation. But I have never seen such TTF files. Usually, its "regular", "bold", "italic" and "bold-italic" faces and corresponding files. A special "underline" face would have to exist in four combinations in order to support underlining any of the other four styles. It's probably bogus to even provision for this case.

Patch-set: 1