Update patch set 1
Patch Set 1:

These are escape sequences for the Terminal. What they do is ask the terminal "please draw the coming text with an underline". The Terminal then asks the app_server "draw the text with an underline". And the app_server must do the drawing (of the text, and the underline).

What's missing is the last part in app_server. It is, essentially, just drawing a line, if we want to do it in the simple way. As Stephan mentionned, we can do a lot more complex things (making sure the line does not touch the characters, having lines of different colors and styles, etc), but it would be nice to have at least a very simple line to start with.

This is something to do on the app_server side, where graphics rendering is done. So the escape sequences will not help, they are not some kind of magic that makes the computer draw underlined text without writing any code.

Patch-set: 1
Label: Code-Review=0