Update patch set 1
Patch Set 1:

The actual drawing of glyphs in app_server happens here:

https://git.haiku-os.org/haiku/tree/src/servers/app/drawing/Painter/AGGTextRenderer.cpp

You can see on line #287 and following (inside the debugging #ifdef), how to draw something simple such as a line (the disabled debugging code draws a bounding box around the glyph).

So all that is missing for a very simple line is to get the information that an underline is to be drawn to that function. I am not sure if that info should be stored in the FontCacheEntry. Maybe it should instead be a field of the drawing context, and as such on the Painter object (which duplicates those fields for no good reasons).

Patch-set: 1
Label: Code-Review=0