Update patch set 1
Patch Set 1:

Is waiting 64 seconds here really ok? I suspect we are trying to send things before network is completly up, and if that happens I think the retry should be in the sub second range? The more interesting thing would be why DHCP tries to do things before network is considered up. My change was just to retry a lot faster as a quick workaround and this maybe is going back to a very long timout. The network code has a changed a lot since I looked at it though...

Patch-set: 1