Update patch set 1
Patch Set 1:

To be clear, there are two delays involved. There is a retry delay (starts at 500ms, then doubles at every try IIRC). And there is a timeout, after which we consider the DHCP negociation has failed, and we start over completely.

This only changes the second thing, that is, we will actually retry using the delay now, for up to 64 seconds, re-sending the DHCP discover message only.

Before this change, the process would immediately timeout, and we would start over, sending a burst of DHCP discover and bogus requests until it somehow allowed us to get a lease from a server.

If we want to retry quicker, the other timeout, the 500ms and increasing one, should be adjusted. But we should still retry the initial discover message for some time (here, 64 seconds) before giving up and going with a self assigned IP in the 169.x range (which is the fallback when no DHCP is found).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0