Update patch set 8
Patch Set 8:

> Patch Set 4:
> 
> Umm I am not familiar with the standard procedure here so please pardon my ignorance, but when does a patch get merged after a code review?

We are trying to be very careful with reviews, so usually the process is that one developers review your changes and gives a +2 if everything looks fine, and a second dev reads it again and merge if he also sees nothing to comment about.

As you can see, I did not spot anything, but Korli did in the second review.

The long term plan is to also have bots checking that the code builds, and that it does not introduce coding style violations, but we do not have this in place yet.

Don't worry, it's ok for patches to linger in Gerrit for some time, usually we are not in a hurry to get things merged. You can work on other things meanwhile.

Patch-set: 8