Update patch set 1
Patch Set 1:

It seems pretty mechanical, as far as a change goes (the original work to make a Codec Kit.) Can anyone point me to where I can read about the rationale for this move? Is it because much of this was previously private to the Media Kit and we needed it exposed for WebPositive and other apps?

Patch-set: 1