Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It seems pretty mechanical, as far as a change goes (the original work to make a Codec Kit.) Can anyone point me to where I can read about the rationale for this move? Is it because much of this was previously private to the Media Kit and we needed it exposed for WebPositive and other apps?

So, there's at least an app using it (HaikuRadio). But the reason is that I want to decouple reading media files and writing them. (You can see a very similar API in stagefright however)

I think BMediaFile is not designed properly. So my plan was to introduce BMediaStream, which is something very similar to BMediaTrack, that's why I've been saying the API is still unfinished. There are at least two reasons for why BMediaFile isn't designed properly for what we want to do.

The first is that you can't use BMediaFile in r/w mode. So, you will end up having two BMediaFiles, that makes little sense.

The second, is that we need more advanced seeking and notification capabilities. We need a way so that the caller know what's the seek range under the hood (something similar to WebAudio SeekRange), and we need a way for the caller to receive notifications about what's going on. For example, metadata updates, re-buffering and so on. I think you can see how adding those features needs basically a BMediaFile rework, that's going to happen with BMediaWriter, BMediaExtractor, BMediaStreamer and BMediaStream. Doing so, I have WIP a plugin called file_streamer, my aim would be that every app pass between that plugin so that meta data can be kept in sync in the disk.

Patch-set: 1