Update patch set 1
Patch Set 1:

> Oh, and going to specify that...it's not possible by modifying BMediaFile, but it's possible to keep BMediaFile R5 compatible and make it use the newever and more advanced API of the codec kit.

I fail to see how it is "not possible." BMediaFile has 47 (!!!) unused virtual methods; and the vast majority of the class's functions are non-virtual. Modifying the class behavior or adding alternative constructors as was done for BView and descendants for the Layout API seems entirely fine; I don't see what you would need to add that so fundamentally changes the makeup of the class itself.

Patch-set: 1