Update patch set 1
Patch Set 1:

> So, there's at least an app using it (HaikuRadio). But the reason is that I want to decouple reading media files and writing them.

Why?

Windows' Media Foundation APIs technically do this (IMFMediaSource and IMFMediaSink), but this is because those objects are actually graph nodes in their media graph API. So you can wire a MediaSource to a MediaFilter chain, and then to a MediaSink, and thus create an effects pipeline.

But BMediaFile is *not* a BMediaNode. If it were, then I would agree, but it is not. (I think introducing BMediaSource/BMediaSink nodes could make sense, but keeping BMediaFile as the "low-level" file interface that does both reading and writing makes more sense than hiding it behind those like some APIs do.)
 
> The first is that you can't use BMediaFile in r/w mode. So, you will end up having two BMediaFiles, that makes little sense.

Indeed that makes little sense. Why is that an inherent limitation of the API though? Can't it be changed to allow this?

> The second, is that we need more advanced seeking and notification capabilities. We need a way so that the caller know what's the seek range under the hood (something similar to WebAudio SeekRange), and we need a way for the caller to receive notifications about what's going on. For example, metadata updates, re-buffering and so on.

A "streamer" plugin is distinct from a "codec" plugin already, yes? So HTTPStreamer is what fetches data, and the FFmpeg plugin is what streams audio. Now, the FFmpeg plugin may initiate a read, and the HTTPStreamer will then have to do something about that; but fundamentally it's the streamer that is responsible for issuing a "re-buffering" notification. Thus it seems to me that this should just be an extension of a BDataIO -- a BStreamingDataIO or the like, and notifications are issued by it.

Then BMediaTrack's API does not have to change, and the notifications are largely optional; obviously it will be rather annoying if an application does not present them, and it buffers video for 5 seconds with no indication of what is going on. But BMediaTrack is mostly OK as-is, it just needs a minor rework to take BStreamingDataIO or whatever into account, and let the developer set or fetch it.

Patch-set: 1