Update patch set 1
Patch Set 1:

> That will introduce a lot of complexity internally. And anyway, it's not that we can have a different constructor easily, unless some flags are passed.

You are again being vague. I outlined in my previous comment exactly the way I understand this, and how it works in other OSes and APIs, and how I think it would make sense to do it here, involving a BStreamingDataIO or the like, and BMediaTrack changing only marginally at best.

But also reviewing BMediaFile, it seems it's R/W already, and calling the constructor is an implicit Open, and the destructor is an implicit Close. So I still don't know what functions need to be added that are "mandatory calls".

> Both. There's a lot of functionality needed to support streaming servers too...

Hence a BStreamingDataIO that would give these notifications; it should not be BMediaStream's job specifically afaict.

> Not sure what you mean by "moving this locking" can you extend?

You mentioned some sort of locking around streaming in MediaPlayer, I thought you meant this was to make BMediaTrack usable, and was asking why MediaPlayer was doing the locking and not the class itself.

Patch-set: 1