Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > That will introduce a lot of complexity internally. And anyway, it's not that we can have a different constructor easily, unless some flags are passed.
> 
> You are again being vague. I outlined in my previous comment exactly the way I understand this, and how it works in other OSes and APIs, and how I think it would make sense to do it here, involving a BStreamingDataIO or the like, and BMediaTrack changing only marginally at best.

To add all the functionality I listed before, you will end up in a mega-blob still retaining the issues I outlined.

> 
> But also reviewing BMediaFile, it seems it's R/W already, 

No.

>and calling the constructor is an implicit Open, and the destructor is an implicit Close. So >I still don't know what functions need to be added that are "mandatory calls".

No. There's a sniffing process which is separated and should be allowed to happpen at client request. For example when a stream goes down, the caller may want to instantiate a new request, it's not very smart to delete the BMediaFile and instantiate a new one.

> 
> > Both. There's a lot of functionality needed to support streaming servers too...
> 
> Hence a BStreamingDataIO that would give these notifications; it should not be BMediaStream's job specifically afaict.

The BAdapterIO has been designed to be only a low level handle. It does not handle metadata and notifications because those may come from higher levels as well.

> 
> > Not sure what you mean by "moving this locking" can you extend?
> 
> You mentioned some sort of locking around streaming in MediaPlayer, I thought you meant this was to make BMediaTrack usable, and was asking why MediaPlayer was doing the locking and not the class itself.

MediaPlayer locks when opening streams because both the API and the app are not designed with this in mind.

Patch-set: 1