Update patch set 1
Patch Set 1:

> In media2 this stuff will indeed be allowed to be spawned as a media source and a media destination.

Yes, OK, that makes sense, that's what I advocated for *as an option.* I don't see why BMediaFile can't back that solution.

> In fact this is one of the reasons for chosing this design. 

That, I do not see; in fact I don't see why BMediaFile is not already adequate for this exact purpose, i.e. why the Codec Kit is really needed at all.

> > > The first is that you can't use BMediaFile in r/w mode. So, you will end up having two BMediaFiles, that makes little sense.
> > 
> > Indeed that makes little sense. Why is that an inherent limitation of the API though? Can't it be changed to allow this?
> 
> No, you can't. Because most codecs are not designed to work this way and it'd be very inefficient and corruption-prone anyway, especially considering non-perfectly-seekable formats.

I am very confused. First you said that it didn't make sense BMediaFile could *not* be opened in R/W mode. But it can, and now you are saying that it doesn't make sense that it *can*. Further, the vast majority of formats used in the present day are perfectly seekable; and the ones that are not, I presume FFmpeg will throw an error if you attempt to seek while in write mode, yes?

So having it "open in R/W mode" but only allowing "end-of-file" writes is how FFmpeg handles this already (reads can still be done from already-written portions I believe), so blocking R/W opens does not make sense.

> There's BAdapterIO already. But it's responsible only for low level r/w. Higher level functionalities are split between the extractor writer and streamer in a consistent way.

What is a "higher level functionality"?

I would expect that, over a LAN connection of 60MB/s+, HTTP streaming looks very similar to disk streaming, i.e. there is so little latency to make streaming look like reading a file. In that case, I would expect the streaming system to behave nearly identically to the way things currently do, that is, the "streaming logic" is only there to notify users when things are buffering and will be for some time, and for how long, etc.

Indeed there seems to be a need for an "immediately seekable range", as in, cached data. But this can be done in one datatype that defines a start/end time, and thus one additional API call to BMediaFile or BMediaTrack, right? So why do we need an entire new "kit" for this functionality?

> I don't think it will work correctly. We had already a lot of brainstorming at the time of the contract and there are various reasons so that it will just make things more complex.

I guess I don't have as much experience working with and designing media APIs as you, because I cannot think of any reasons. Can you elaborate on what these are?

Patch-set: 1