Update patch set 1
Patch Set 1:

> To add all the functionality I listed before, you will end up in a mega-blob still retaining the issues I outlined.

You listed "notifications" (for buffering information), and "metadata writing". I don't know what other functionality you listed.

> > But also reviewing BMediaFile, it seems it's R/W already, 
> 
> No.

http://xref.plausible.coop/source/xref/haiku/headers/os/media/MediaFile.h#54

"	//	these three constructors are for read-write access"

https://www.haiku-os.org/legacy-docs/bebook/BMediaFile.html on the other hand implies these are not read-write but write only indeed. But I don't see any reason they are? ReadFrames() etc. in BMediaTrack just say that they may return "EPERM" if it wasn't opened for reading. So we can just make that work, if possible, right? Codecs that don't support it can continue returning EPERM.

Or we can indeed add a new constructor that is truly R/W... either way, does not seem so hard.

> >and calling the constructor is an implicit Open, and the destructor is an implicit Close. So >I still don't know what functions need to be added that are "mandatory calls".
> 
> No. There's a sniffing process which is separated and should be allowed to happpen at client request. For example when a stream goes down, the caller may want to instantiate a new request, it's not very smart to delete the BMediaFile and instantiate a new one.

If the stream went down, does that not invalidate the BMediaTracks, and the seek locations, etc. etc. etc.? So how would it *not* make more sense to delete and make a new BMediaFile when restarting streaming? It seems that it would be a mess if you didn't invalidate the BMediaFile upon stream stop.

That is different from a temporary connection failure. In that case, if the HTTPStreamer or whatever can repair the connection before the codecs notice, then again, I don't know why the codecs would ever want to notice. This implies the stream is seekable though; if it isn't seekable, then the only thing to do would be to make a new BMediaFile, as virtually *everything* about the stream will now have changed after losing the connection.

> The BAdapterIO has been designed to be only a low level handle. It does not handle metadata and notifications because those may come from higher levels as well.

I don't expect BAdapterIO to care about metadata, indeed. But I don't know why it ever would; it's just a way to read and write files/data, right? So I do not understand what the problem is here.

> MediaPlayer locks when opening streams because both the API and the app are not designed with this in mind.

OK. Why cannot this locking be moved to a lower level? Or MediaPlayer repaired so that it does support streams more properly?

Patch-set: 1