Update patch set 1
Patch Set 1:

> > Yes, OK, that makes sense, that's what I advocated for *as an option.* I don't see why BMediaFile can't back that solution.
> 
> Because this way you are replicating BFileInterface which never worked properly.

Plenty of stuff in BeOS didn't work properly because it was poorly implemented, and works much better in Haiku. Some stuff works worse because it was used so little we did not implement it well. It "never working properly" is not a reason that it will never work properly.

I really do not see how IMFMediaSource/IMFMediaSink, BFileInterface, and your vague proposals are so different. Of course their APIs have different function names, buffer pull vs. push, etc. are different, but those are inconsequential differences (the last specifically is the fault of the node graph framework, not the class, if it is a problem at all.)

If you are saying that BFileInterface is in some fundamental way flawed in its design and not only implementation, and what you are proposing is indeed substantially different and will not have this flaw, of course I want to hear it. But I don't see where you have said anything about this at all.

> > So having it "open in R/W mode" but only allowing "end-of-file" writes is how FFmpeg handles this already (reads can still be done from already-written portions I believe), so blocking R/W opens does not make sense.
> 
> It'd probably break ABI.

The comment in the header indicates that it already opens in R/W mode. So is this false? And as I said above, I don't see how it would break ABI; the functions like ReadFrames() which used to return B_NOT_ALLOWED under W/O mode can now return the actual frames and B_OK. How does this break ABI at all?

> > > There's BAdapterIO already. But it's responsible only for low level r/w. Higher level functionalities are split between the extractor writer and streamer in a consistent way.
> > 
> > What is a "higher level functionality"?
> 
> All that comes from the streamer-reader-demuxer-decoder chain

The "streamer" is BAdapterIO; the rest of the chain is found in our plugin system already. Again, buffering information should from BAdapterIO, as only it knows network latency, estimated download time, etc.

What other information is there to pass back that we cannot do without substantially modifying the API? Adding a single call to BAdapterIO, or 2-3 to BMediaFile/BMediaTrack sounds OK with me; much better than adding an entire new interface.

> You can't assume fast connections by default it's just going to raise a lot of complaints

Of course, hence why we do need some system to notify when buffering is occurring and the "seekable range" as you said. I see any number of ways that this can be integrated with the current BMediaFile and BAdapterIO. I don't see any reason this warrants an entirely new kit.

> > Indeed there seems to be a need for an "immediately seekable range", as in, cached data. But this can be done in one datatype that defines a start/end time, and thus one additional API call to BMediaFile or BMediaTrack, right? So why do we need an entire new "kit" for this functionality?
> 
> There's need of notifications...

So add a SetTarget method to BMediaTrack, and have it or anything else up the stack send BMessages to said target. Or add a SetCallback method and have it call a method directly and bypass the BMessage system. Either way, this does not merit an entirely new kit.

Patch-set: 1