Update patch set 1
Patch Set 1:

> > You listed "notifications" (for buffering information), and "metadata writing". I don't know what other functionality you listed.
> 
> Mouse tracking, making easy to convert files for example.

Is there anything that needs mouse tracking besides Blu-Ray/DVD playback? If not, I would consider treating that as a special case. (Wasn't there some app that did this with BParameterWeb? You can already manipulate a lot of real-time parameters with that, passing through mouse information seems strange but not out of the question.)

I don't know what improvements to file conversion you are speaking of which would, again, require a new API. Obviously MediaConverter does not work so well, but it certainly did on BeOS, so that seems to imply our implementations of things are bad.

> I'd not bet on that. It'd probably break things up. The point is, why hack a class that has not been designed to work this way?

What does "it'd probably break things up" mean?

I don't understand how it's "a hack". There is a function called ReadFrames, and a function called WriteFrames. Previously only one worked at a time and the other returned errors. How is it a hack to make both work at once under "write" mode? That would make perfect sense to me; in fact I think it's how some of the other media toolkits operate (MLT?)

> > Or we can indeed add a new constructor that is truly R/W... either way, does not seem so hard.
> 
> Implement it...

I said that because I truly don't believe it to be hard. You seem to have taken a lot of time and effort working on a new API because you believe that to be easier than fixing the old system. I don't know why, and I stated this hoping you had an answer?

> > If the stream went down, does that not invalidate the BMediaTracks, and the seek locations, etc. etc. etc.? So how would it *not* make more sense to delete and make a new BMediaFile when restarting streaming? It seems that it would be a mess if you didn't invalidate the BMediaFile upon stream stop.
> 
> Not necessarily, you can re-Open a stream and continue writing for example. Depending on the streamer it may even support restoring the status.

OK. In that case, why does the programmer even need to know anything? If this can all be done automatically, shouldn't it be the plugin that makes the decision as to whether or not it's safe to do so, not the third-party programmer? And if it can't or it is not safe, then indeed an error should be returned and a new BMediaFile created. Otherwise the re-open and retires should all happen transparently and the programmer should need to do nothing special at all.

> > That is different from a temporary connection failure. In that case, if the HTTPStreamer or whatever can repair the connection before the codecs notice, then again, I don't know why the codecs would ever want to notice.
> 
> Because apps *needs* to know when something like that happen.

Like what, a connection drop/restart? Then send them through a message or a callback notification...

>  This implies the stream is seekable though; if it isn't seekable, then the only thing to do would be to make a new BMediaFile, as virtually *everything* about the stream will now have changed after losing the connection.
> 
> It's the opposite. Non seekable stuff can just be restored in a peaceful way.

OK. Then we should perhaps send a notification, but there is no reason to force the programmer to do any API calls to restart the stream, if we can automatically do so. If we can't, then intervention is probably required and thus a new BMediaFile should be made as there are almost certainly new parameters of some kind.

> > OK. Why cannot this locking be moved to a lower level? Or MediaPlayer repaired so that it does support streams more properly?
> 
> Apps want granular control about what happens, you could just spawn a new thread and instantiate the BMediaFile but it's not going to be the easiest way and you'd end up with more threads than what are needed.

Presumably the app already has a "data processing thread" (it had to instantiate the BMediaFile in the first place, right?) so it can just re-instantiate it there.

Patch-set: 1