Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> It feels like we aren't even discussing the merits of these codec kit changes. From my first question Dario mentioned BMediaStream, which I do see in these changes, but why does adding that require pulling so many things into a new Codec Kit? To me it seems like BMediaStream could just be added to the Media Kit, if we really need it, which maybe we don't.

BMediaStreamer is there, not BMediaStream...

> 
> If there is not a really good reason for creating a new Codec Kit, I would rather we not.
> 
> Also in general Dario I am not sure you understand how we really want to avoid major code changes to kits for Haiku R1. It seems like a lot of what you do should maybe wait until R2 is under development, and then maybe not even then.

This is R2 work...not meant to be R1 stuff.

> 
> Breaking things out into a new Codec Kit seems to be making an already complicated Haiku media framework even more complicated.

To support all the stuff outlined before you are not going to succeed with BMediaFile alone.
Handling network latencies for example it's just impossible. I already outlined a lot of stuff that means basically redesigning BMediaFile. What I add to this re-design is that I decouple extractors and writers.

> 
> When I write code I like to ask "what is the smallest possible change I can make", whether I am fixing a bug, or adding a feature. I think that rule really applies to Haiku too. It seems like you don't try to follow that rule.

OK. Apply this patch please. Don't take this bad or a complaint, but if we go with this mindset in R6 there will still be the media_kit and there will be still the same problems I outlined before.

Patch-set: 1