Update patch set 1
Patch Set 1:

> > So add a SetTarget method to BMediaTrack, and have it or anything else up the stack send BMessages to said target. Or add a SetCallback method and have it call a method directly and bypass the BMessage system. Either way, this does not merit an entirely new kit.
> 
> Okay.

Yes if you look at it in the optics of replicating BMediaFile of course we don't need a new kit. But if you take patience and see all the features I listed, you can see there's need of a lower level and powerful kit. Just as I told Ryan, if you don't see the usefulness of this, while keeping the simple BMediaFile still available, I can't do more to convince you. Your replies are always in the optic of BMediaFile, while the kit is much more than that.

So please revert this work and let it be.

Patch-set: 1