Update patch set 1
Patch Set 1:

> > Plenty of stuff in BeOS didn't work properly because it was poorly implemented, and works much better in Haiku. Some stuff works worse because it was used so little we did not implement it well. It "never working properly" is not a reason that it will never work properly.
> > 
> > I really do not see how IMFMediaSource/IMFMediaSink, BFileInterface, and your vague proposals are so different. Of course their APIs have different function names, buffer pull vs. push, etc. are different, but those are inconsequential differences (the last specifically is the fault of the node graph framework, not the class, if it is a problem at all.)
> > 
> > If you are saying that BFileInterface is in some fundamental way flawed in its design and not only implementation, and what you are proposing is indeed substantially different and will not have this flaw, of course I want to hear it. But I don't see where you have said anything about this at all.
> 
> It never worked properly even in BeOS.

Please re-read my first and second paragraphs quoted above: even if it does not work in BeOS, that means not very much about its actual worth, as we have implemented quite a lot of stuff in Haiku that did not work very well in BeOS either.

Further, FFmpeg and codec APIs have changed a lot in the last two decades since R5 came out. So just because it did not work properly in BeOS is not a reason to condemn it because of that also.

So unless you have an actual objection to its API, and an explanation of how your new API does not have this problem, "it does not work" is NOT an argument in the slightest.

> > What other information is there to pass back that we cannot do without substantially modifying the API? Adding a single call to BAdapterIO, or 2-3 to BMediaFile/BMediaTrack sounds OK with me; much better than adding an entire new interface.
> 
> Normally, you are not supposed to instantate a BAdapterIO yourself. And as explained in the mailing list discussion, there is a lot of stuff that needs special support and lower level functionalities that can't be only put in BMediaFile. Like audio over ip for example. 

How does audio over IP need special support, precisely? You still need to read and write frames; only now you have a hard demand on time constraints. Hopefully whoever instantiated the object knows that and will keep up (or notifications can be sent on low buffer counts, just like in USB isochronous transfers or the multi_audio API?)

> > > You can't assume fast connections by default it's just going to raise a lot of complaints
> > 
> > Of course, hence why we do need some system to notify when buffering is occurring and the "seekable range" as you said. I see any number of ways that this can be integrated with the current BMediaFile and BAdapterIO. I don't see any reason this warrants an entirely new kit.
> 
> You can, revert this patch and implement it your way, I'm fine with that.

The whole point of code review is to suggest better ways to do something, and then the person who submitted the original patch comment on them. Saying "you can do it your way" means you are resigning responsibility and abandoning this patch. Is that what you intend here?

But that also does not make much sense. Are you admitting my way is better? Or just that you want to be done arguing about it? This is not how code review usually works. I have some patches open now (e.g. kernel select changes) which collected comments that essentially amount to a complete overhaul of the change (and it's a rather large one.) But I agree it's a much better way of doing things, and I should clean it up and do that change (it wasn't even mine originally, I adopted it.)

> There are a number of reasons for why a more powerful kit is needed. I outlined a lot there, but there are a lot that I'm probably forgiving right now.

The ones you outlined, I made counterpoints to that you have mostly not responded to, and instead replied in a few cases with "Well, implement it then." This means "I don't think it's possible, so prove it to me", giving no reasons for this belief. This is not how developers collaborate.

> > > > Indeed there seems to be a need for an "immediately seekable range", as in, cached data. But this can be done in one datatype that defines a start/end time, and thus one additional API call to BMediaFile or BMediaTrack, right? So why do we need an entire new "kit" for this functionality?
> 
> You are not probably going to support streaming servers with BMediaFile alone. There's just a lot of stuff that needs to be supported.

OK. What else have I missed?

> > So add a SetTarget method to BMediaTrack, and have it or anything else up the stack send BMessages to said target. Or add a SetCallback method and have it call a method directly and bypass the BMessage system. Either way, this does not merit an entirely new kit.
> 
> Okay.

Well, I am glad there is one thing we agree on at least. :)

Patch-set: 1